refactor: 레이아웃 관련 스토어 파일 위치 수정
[#382]
