+++ frontend/src/hooks/google-maps/useCalculatedMapDelta.ts
@@ -1,8 +1,11 @@
 import { useExternalValue } from '@utils/external-state';
 import { getDisplayPosition } from '@utils/google-maps';
 
-import { browserWidthStore, navigatorAccordionWidthStore } from '@stores/componentWidthStore';
 import { getGoogleMapStore } from '@stores/google-maps/googleMapStore';
+import {
+  browserWidthStore,
+  navigatorAccordionWidthStore,
+} from '@stores/layout/componentWidthStore';
 
 export const useCalculatedMapDelta = () => {
   const navigatorAccordionWidth = useExternalValue(navigatorAccordionWidthStore);
