+++ frontend/src/stores/layout/componentWidthStore.ts
