+++ frontend/src/stores/layout/forceOpenAccordionPanelStore.ts
